Refactor Toolkit to reside in agent_toolkits (#1328)
IMO, it's a bit cleaner to have toolkits in `agent_toolkits`.
Previously, a `Toolkit` could have imported from `agent_toolkit` and a
`create_*_agent` could have imported from `Toolkit`.
example: `OpenAPIToolkit` needed `create_json_agent` and
`create_openapi_agent` needed `OpenAPIToolkit`
* Delete unnecessary toolkits (toolkits with only one tool)
* Move `BaseToolkit` to agent_toolkits
* Move toolkits to agent_toolkits
* update notebooks
