+++ langchain/agents/agent_toolkits/__init__.py
@@ -1,7 +1,17 @@
 """Agent toolkits."""
 
 from langchain.agents.agent_toolkits.json.base import create_json_agent
+from langchain.agents.agent_toolkits.json.toolkit import JsonToolkit
 from langchain.agents.agent_toolkits.openapi.base import create_openapi_agent
+from langchain.agents.agent_toolkits.openapi.toolkit import OpenAPIToolkit
 from langchain.agents.agent_toolkits.sql.base import create_sql_agent
+from langchain.agents.agent_toolkits.sql.toolkit import SQLDatabaseToolkit
 
-__all__ = ["create_json_agent", "create_sql_agent", "create_openapi_agent"]
+__all__ = [
+    "create_json_agent",
+    "create_sql_agent",
+    "create_openapi_agent",
+    "JsonToolkit",
+    "SQLDatabaseToolkit",
+    "OpenAPIToolkit",
+]
