+++ langchain/agents/agent_toolkits/openapi/base.py
@@ -6,12 +6,12 @@ from langchain.agents.agent_toolkits.openapi.prompt import (
     OPENAPI_PREFIX,
     OPENAPI_SUFFIX,
 )
+from langchain.agents.agent_toolkits.openapi.toolkit import OpenAPIToolkit
 from langchain.agents.mrkl.base import ZeroShotAgent
 from langchain.agents.mrkl.prompt import FORMAT_INSTRUCTIONS
 from langchain.callbacks.base import BaseCallbackManager
 from langchain.chains.llm import LLMChain
 from langchain.llms.base import BaseLLM
-from langchain.tools.openapi.toolkit import OpenAPIToolkit
 
 
 def create_openapi_agent(
