+++ langchain/agents/agent_toolkits/openapi/prompt.py
@@ -21,3 +21,9 @@ OPENAPI_SUFFIX = """Begin!"
 Question: {input}
 Thought: I should explore the spec to find the base url for the API.
 {agent_scratchpad}"""
+
+DESCRIPTION = """Can be used to answer questions about the openapi spec for the API. Always use this tool before trying to make a request. 
+Example inputs to this tool: 
+    'What are the required query parameters for a GET request to the /bar endpoint?`
+    'What are the required parameters in the request body for a POST request to the /foo endpoint?'
+Always give this tool a specific question."""
