--- langchain/tools/openapi/prompt.py
@@ -1,7 +1,0 @@
-# flake8: noqa
-
-DESCRIPTION = """Can be used to answer questions about the openapi spec for the API. Always use this tool before trying to make a request. 
-Example inputs to this tool: 
-    'What are the required query parameters for a GET request to the /bar endpoint?`
-    'What are the required parameters in the request body for a POST request to the /foo endpoint?'
-Always give this tool a specific question."""
